Update processing ux (#6)
* Refactor to remove commands and make format consistent

* Add correct requirements

* Fix use of date format in readme

* Add ignore_existing flag

* Simplify instructions

* Refactor viewer

* Update use of date in cli

* Update images to match new api

* Apply black

* Update docs

* Remove tests for 3.7, update black

* Fix python version

* Apply black on init files

* Install other dependencies

+++ wikicat/processing/merge_tables.py
@@ -10,7 +10,7 @@ python -m wikicat.processing.merge_tables \
     --year 2018 \
     --month 12 \
     --day 20 \
-    --base_dir /path/to/save/intermediate/files
+    --base_dir "~/.wikicat_data/"
 """
 import argparse
 import json
@@ -18,61 +18,85 @@ from pathlib import Path
 
 import pandas as pd
 
+
 def merge_tables(page_csv_filepath, category_csv_filepath):
-    page_df = pd.read_csv(page_csv_filepath, \
-            usecols = ['page_id', 'page_title', 'page_namespace'], \
-            dtype={'page_id': 'int32', 'page_title': str, \
-                'page_namespace': 'int32'}).set_index('page_id')
+    """
+    Merge the page and category tables into a single table.
 
-    category_df = pd.read_csv(category_csv_filepath).set_index('cl_from')
+    Parameters
+    ----------
+    page_csv_filepath : str
+        Path to the page CSV file.
+    category_csv_filepath : str
+        Path to the category CSV file.
 
-    full_df = pd.merge(page_df, category_df, left_index=True, \
-            right_index=True, how="right", copy=False)
+    Returns
+    -------
+    pandas.DataFrame
+        The merged table.
+
+    Notes
+    -----
+    This step may take a while to run (1h+). It is recommended to run this
+    script on a machine with a lot of RAM.
+
+    Example
+    -------
+    >>> page_csv_filepath = "~/.wikicat_data/enwiki_2018_12_20/page.csv"
+    >>> category_csv_filepath = "~/.wikicat_data/enwiki_2018_12_20/categorylinks.csv"
+    >>> df = merge_tables(page_csv_filepath, category_csv_filepath)
+    >>> print(df.head(10))
+    >>> df.to_csv("~/.wikicat_data/enwiki_2018_12_20/full_catgraph.csv")
+    """
+    page_df = pd.read_csv(
+        page_csv_filepath,
+        usecols=["page_id", "page_title", "page_namespace"],
+        dtype={"page_id": "int32", "page_title": str, "page_namespace": "int32"},
+    ).set_index("page_id")
+
+    category_df = pd.read_csv(category_csv_filepath).set_index("cl_from")
+
+    full_df = pd.merge(
+        page_df, category_df, left_index=True, right_index=True, how="right", copy=False
+    )
 
     # keep only certain columns
-    full_df = full_df[["page_namespace", "page_title", "cl_to", \
-        "cl_type"]]
-    
-    # keep only articles and categories
-    full_df = full_df[(full_df['page_namespace'] == 0) \
-            | (full_df['page_namespace'] == 14)]
+    full_df = full_df[["page_namespace", "page_title", "cl_to", "cl_type"]]
 
-    full_df.index.name = 'page_id'
+    # keep only articles and categories
+    full_df = full_df[
+        (full_df["page_namespace"] == 0) | (full_df["page_namespace"] == 14)
+    ]
+
+    full_df.index.name = "page_id"
 
     full_df.sort_index(inplace=True)
 
     return full_df
 
-def main(
-        year,
-        month,
-        day,
-        base_dir,
-        page_csv_load_path,
-        category_csv_load_path,
-        save_filepath,
-    ):
+
+def main(year, month, day, base_dir, ignore_existing):
     base_dir = Path(base_dir).expanduser()
     base_dir.mkdir(parents=True, exist_ok=True)
 
-    int_dir = base_dir / f"enwiki_{year}{month:02d}{day:02d}"
-    int_dir.mkdir(exist_ok=True)
+    int_dir = base_dir / f"enwiki_{year}_{month:02d}_{day:02d}"
+
+    if not int_dir.is_dir() and not ignore_existing:
+        raise ValueError(
+            f"Intermediate directory {int_dir} does not exist. Make sure to run the download_dump.py script first."
+        )
 
     # default names for filepath arguments
-    if page_csv_load_path == "None":
-        page_csv_load_path = int_dir / f"page.csv"
-    
-    if category_csv_load_path == "None":
-        category_csv_load_path = int_dir / f"categorylinks.csv"
-    
-    if save_filepath == "None":
-        save_filepath = int_dir / f"full_catgraph.csv"
+    page_csv_load_path = int_dir / "page.csv"
+    category_csv_load_path = int_dir / "categorylinks.csv"
+    save_filepath = int_dir / "full_catgraph.csv"
 
     # merge the tables
     full_df = merge_tables(page_csv_load_path, category_csv_load_path)
 
     # Save the full dataframe to a CSV file
     full_df.to_csv(save_filepath)
+
 
 def parse_args():
     parser = argparse.ArgumentParser(
@@ -85,32 +109,22 @@ def parse_args():
         formatter_class=argparse.ArgumentDefaultsHelpFormatter,
     )
 
-    parser.add_argument("--year", type=int, required=True, help="Year of the dump")
-    parser.add_argument("--month", type=int, required=True, help="Month of the year")
-    parser.add_argument("--day", type=int, required=True, help="Day of the month")
+    parser.add_argument("--year", "-y", type=int, required=True, help="Year of dump")
+    parser.add_argument("--month", "-m", type=int, required=True, help="Month of dump")
+    parser.add_argument("--day", "-d", type=int, required=True, help="Day of dump")
 
     parser.add_argument(
-        "--base_dir", type=str, help="Base directory for intermediate files", default="~/.wikicat_data"
+        "--base_dir",
+        type=str,
+        help="The directory containing a directory in the form enwiki_<YYYY>_<MM>_<DD>. The latter contains the dumps and your output file will be stored there.",
+        default="~/.wikicat_data",
+    )
+    parser.add_argument(
+        "--ignore_existing",
+        action="store_true",
+        help="Ignore cached output file. Only do this if you previous generated the file and want to regenerate it.",
     )
 
-    parser.add_argument(
-        "--page_csv_load_path",
-        type=str,
-        help="Full path (including filename) to the page CSV file (overrides yyyy/mm/dd args)",
-        default="None",
-    )
-    parser.add_argument(
-        "--category_csv_load_path",
-        type=str,
-        help="Full path (including filename) to the category CSV file (overrides yyyy/mm/dd args)",
-        default="None",
-    )
-    parser.add_argument(
-        "--save_filepath",
-        type=str,
-        help="Full path (including filename) to save the merged CSV file to (defaults to {intermediate_dir}/full_catgraph.json)",
-        default="None",
-    )
     return parser.parse_args()
